Created new log utils and replaced old one
